Moved useEffect for clearing cart to PaymentDetails component so that payment-success page would stay server-side
