+++ src/components/PaymentDetails.jsx
@@ -1,6 +1,7 @@
 "use client"
 import { useSearchParams } from 'next/navigation'
 import { useEffect, useState } from 'react'
+import { useCart } from "@/app/context/CartContext"
 import OrderSummary from "./OrderSummary"
 import calculateTotal from "../../lib/calculateTotal"
 
@@ -9,6 +10,17 @@ export default function PaymentDetails({ products }) {
     const payment_intent = searchParams.get('payment_intent')
     const [paymentDetails, setPaymentDetails] = useState(null)
     const [total, setTotal] = useState(0)
+
+    const { clearCart } = useCart()
+
+    useEffect(() => {
+        const alreadyCleared = localStorage.getItem("cartCleared")
+
+        if (!alreadyCleared) {
+            clearCart()
+            localStorage.setItem("cartCleared", "true")
+        }
+    }, [])
 
     useEffect(() => {
         if (payment_intent) {
